boards/xtensa/esp32: Rename the common files as esp32_board_* to avoid any naming collisions.
